requirements.txt
fastapi==0.110.3
psycopg2-binary==2.9.9
uvicorn==0.29.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2
pytest~=8.2.1
Faker~=25.3.0
pydantic~=2.7.1
retrying==1.3.4
requests~=2.32.3
urllib3~=2.2.1
matplotlib~=3.9.0
numpy~=1.26.4
redis~=5.0.8